AdaDelta warmup for the conparser.  Motivation: AdaDelta results in higher scores overall, but learns 0s for the weights of the pattn and lattn layers.  AdamW learns weights for pattn, and the models are more accurate than models trained without pattn using AdamW, but the models are lower scores overall than the AdaDelta models.
This improves that by first running AdaDelta, then switching.

Now, if --multistage is set, run AdaDelta for half the epochs with no
pattn or lattn.  Then start the specified optimizer for the rest of
the time with the full model.  If pattn and lattn are both present,
the model is 1/2 no attn, 1/4 pattn, 1/4 pattn and lattn

Included is a test of both the two and three step versions

Also add a false version of the option - we may very well want to make
--multistage the default